Append plugin api version to Bazel plugin version.
This CL adds the major version of the plugin api to the version of the plugin compatible with it. This will enable releasing different versions of the plugin with the same release each supporting different IDE version.

For a release on 2021.07.14 targeting IDE build 211.7628.21, the plugin version will be "2021.07.14-api-version-211".

PiperOrigin-RevId: 384871619