ModuleEditorImpl: Check the value of an option before setting it to work around a slow operation
With the new project model, Module#setOption can be slow when called on
existing modules as that method kicks of a series of operations (e.g.
notifies listeners about updates). If a high number of modules is
involved (-> on large projects), that slowness expresses itself as
freezes to users as ModuleEditorImpl is called on the EDT.

From 2021.2.1 on, the implementation of Module#setOption will check
whether the option already has the passed value (see [IDEA-274168](https://youtrack.jetbrains.com/issue/IDEA-274168#focus=Comments-27-5086318.0-0)).
Until then, we include that check in ModuleEditorImpl.

PiperOrigin-RevId: 387831950