Fix NPE in BlazeXmlToTestEventsConverter#parseComparisonData
The message can be null as e.g. null checks in
BlazeXmlToTestEventsConverter#pruneErrorMessage show.

We additionally checked other usages of ErrorOrFailureOrSkipped.message
and added an additional null check for safety in
BlazeXmlToTestEventsConverter#processTestCase. This might not have been
necessary and is unrelated to the reported NPE but could have
represented another potential NPE.

Fixes #2825

PiperOrigin-RevId: 389914193