Don't use IDE selected device if instrumentation test has target_device
Currently, android_instrumentation_test targets will always run with a
device selected through the IDE.  Some time ago a `target_device`
attribute was added to android_instrumentation_test rule that manually
specifies the device the test should target.  In this case, the IDE
should not force the test to run on the device of its choosing.

This CL adds detection for the `target_device` attribute and skips
device specification flags if the attribute exists.

PiperOrigin-RevId: 390412767