Fix a memory leak by properly hooking up Blaze's console view with disposers
Even though ConsoleView is a Disposable and needs to be disposed to clean up all of its open resources, we didn't connect it properly to a parent disposable all the way up the chain.

In the new implementation of Blaze's tool window, there are even multiple console views, making this issue even more severe. That new implementation unfortunately doesn't follow a strict separation between view and business logic. As instances of ConsoleView are created in TasksTreeConsoleBehaviour, we need to pass down a parent Disposable from the overall project service. If instead the ConsoleView instances had been created in a view class (as a clear separation would dictate), we would have simply registered all relevant views as being disposable, with the tool window content being the highest-level parent disposable.

PiperOrigin-RevId: 399117020