Add Kotlin example project
This CL adds examples directory to the plugin code and adds a simple example for Kotling project. The examples directory is excluded from the project views of the plugin code.

This CL also updates the rules_kotlin version in workspace.oss to the latest 1.3. It also adds dependency `diffutils` because without it failing tests that use Google truth crash with java.lang.NoClassDefFoundError: `difflib/DiffUtils` instead of reporting the textual difference.

PiperOrigin-RevId: 407541490