Add OptionMetadataTag.EXPLICIT_IN_OUTPUT_PATH
See OptionMetadataTag.java for documentation on what the tag does.

Also audited the transitive calls of OutputDirectories.buildMnemonic and annotated options that seemingly can safely have the tag.

(Note that some, like AppleConfiguration, have a getOutputDirectory that is too intricate for me to comfortably add the tag to any of the associated options.)

Related to https://github.com/bazelbuild/bazel/issues/14023.

PiperOrigin-RevId: 410335434