Report identified errors/warnings for some uncovered cases also in the new problems view
Not all cases were covered in the new problems view so far. There were some code paths which called the IssueOutputFilter (= handler for the old problems view) but didn't call the ToolWindowTaskIssueOutputFilter (= handler for the new problems view). Those code paths were the ones which didn't require navigation from the problems view to the Blaze console. This CL adds this connection and additionally brings the code into better shape.

PiperOrigin-RevId: 433682111