Add SyncStrategy to BuildSystem.
This indicates the approach to be taken during sync builds: whether or not to parallelize the builds. Use the SERIAL strategy for Bazel since plain Bazel does not support parallism.

Also update the test utility code BuildSystemProviderWrapper to allow this to be overridden in test code.

This will be integrated into the sync code in subsequent CLs. The new code is not used yet.

PiperOrigin-RevId: 436189354