Always set the BuildBinaryType on UpdatePhaseTask.
UpdatePhaseTask is an @AutoValue class, and so it's builder requires a BuildBinaryType to be set; failing to do do will throw a runtime exception. Remove this potential crash by falling back to BuildBinaryType.NONE in the case that there are no build phase stats in the build result.

I've not seen this failing in this way outside of broken tests, but removing a codepath that will always crash can't be a bad thing.

PiperOrigin-RevId: 438544053