Use BlazeInfo returned from a BuildInvoker to ensure it's appropriate for the context.
Now that we decide later in the sync process which type of biujld invoker we're using, we can't run `blaze info` from ProjectStateSyncTask and assume that we can use the results later, as we maybe started using a remote invoker which will return different info. Instead, move that logic inside the BuildInvoker, and then keep a reference to the invoker used in BlazeSyncBuildResult.

In ProjectUpdateSyncTask.runProjectUpdatePhase(), require callers to pass a BlazeInfo in since it's no longer availble in SyncProjectState. There two callers: one in the main sync flow, where we just pass in the info from the invoker used for thay sync. The other used from CleanProjectTargetsSyncAction and in that case we used the BlazeInfo persisted fro the last full sync performed. This is a behaviour change (and likely a subtle bug fix) as we ensure that we used the BlazeInfo corresponding to the last sync we did, rather then the current sync settings (which could be different).

PiperOrigin-RevId: 441732372