Renaming build shard nodes to avoid ordering problems
The build shard nodes in the tree view of the new blaze console are renamed to avoid UI issues w.r.t. numbering/ordering. At the time of creating a build shard node, a random 8-character long alphanumeric string (generated from a UUID) is appended to the name of a build shard task (instead of a shard number).

PiperOrigin-RevId: 444478691