Pool strings during BEP parsing
Adds a new scope, SharedStringPoolScope, for storing shared references to Strings to conserve memory, as BEP protos often contain a high number of duplicate strings, particularly during sharded builds. Methods in BuildResultHelper and ParsedBepOutput are overloaded to receive an optional String interner. If the interner is not passed as an argument, the parser creates a local interner for reusing strings.

When the BEP emits a NamedFileSet (and a string pool is available), it is replaced with a copy whose string references are interned, combining references for strings seen within a single parsing invocation and across concurrent parsing operations with in a single sync.

PiperOrigin-RevId: 451425350