Fix blaze query used to find targets for a Kotlin source file
Blaze's `same_pkg_direct_rdeps` is unable to detect targets that are expanded by Kotlin macros. The solution is to find the enclosing package of the source file and then recursively expand targets within it using `rdeps(..,, 2)`. This fix has been constrained to Kotlin files (as issue is specific to Kotlin macros) and partial sync operation (as it only affects single source files, Full sync uses the BUILD file so does not have this issue).

PiperOrigin-RevId: 452355124