Create entrypoint for the new sync prototype.
Using an entrypoint for this allows all of the sync code to be kept entirely seperate, and makes it easy to exclude all the new code from the build to ensure it doesn't add any risk when not in use. Move the existing experiment definition behind this new EP interface to keep that encapsulated.

(cherry picked from commit a3053fb6694fc73bd1a9fecf97e0b21699aaef8e)