instrumentation info parser: throw exception rather than logging to context
InstrumentationInfo.getInstrumentationInfo uses the blaze context only to
log errors. This makes it harder to reuse it in situations where we don't
have a blaze context and console. This CL changes the method to throw
an exception with an error message rather than dumping the error to the
console, which allows future CLs to use this method from places that
don't have a context.

(cherry picked from commit 7dbf09f6363d2def8aa0aca6304914f409afbe4b)