Delete ListenableFuture from the plugin jar in the OSS github build.
Since commit b3c2b00a3, the plugin now has a dependency on guava. Since the code in question there is intended to be independant of the plugin API, we cannot rely on the version of guava that is included in IntelliJ, so must depend on guava directly instead. This results in the guava classes being included in the plugin jar, which in general is not a problem, except when a guava class is used in an interface between plugins and the IJ platform. ListenableFuture is used in this way, and is the only known guava class that is. Removing it from the plugin jar should be safe, since the ListenableFuture interface seems very unlikely ever to change.

Note, this may lead to problems if the version of guava we depend on differs in important ways from the version embedded in the IJ API. Solving this potential problem is a much larger task that is not a priority right now.

This should resolve https://github.com/bazelbuild/intellij/issues/4264.

(cherry picked from commit 07d622caa058974fb3ee16a60924fea907e0efd3)